fix(template): unduplicate render_template code
The filters was missing from the types.Jinja2 type.

This change factorizes the code to share the rendering method.
