+++ mergify_engine/actions/dismiss_reviews.py
@@ -19,6 +19,7 @@ import voluptuous
 from mergify_engine import actions
 from mergify_engine import check_api
 from mergify_engine import context
+from mergify_engine import jinja2_utils
 from mergify_engine import rules
 from mergify_engine import signals
 from mergify_engine.clients import http
@@ -61,10 +62,10 @@ class DismissReviewsAction(actions.Action):
                 )
 
             try:
-                message = await ctxt.pull_request.render_template(
-                    self.config["message"]
+                message = await jinja2_utils.render_template(
+                    ctxt.pull_request, self.config["message"]
                 )
-            except context.RenderTemplateFailure as rmf:
+            except jinja2_utils.RenderTemplateFailure as rmf:
                 return check_api.Result(
                     check_api.Conclusion.FAILURE,
                     "Invalid dismiss reviews message",
