+++ mergify_engine/actions/post_check.py
@@ -21,6 +21,7 @@ import voluptuous
 from mergify_engine import actions
 from mergify_engine import check_api
 from mergify_engine import context
+from mergify_engine import jinja2_utils
 from mergify_engine import rules
 from mergify_engine import signals
 from mergify_engine.dashboard import subscription
@@ -78,11 +79,12 @@ class PostCheckAction(actions.Action):
             "check_conditions": rule.conditions.get_summary(),
         }
         try:
-            title = await ctxt.pull_request.render_template(
+            title = await jinja2_utils.render_template(
+                ctxt.pull_request,
                 self.config["title"],
                 extra_variables,
             )
-        except context.RenderTemplateFailure as rmf:
+        except jinja2_utils.RenderTemplateFailure as rmf:
             return check_api.Result(
                 check_api.Conclusion.FAILURE,
                 "Invalid title template",
@@ -90,10 +92,10 @@ class PostCheckAction(actions.Action):
             )
 
         try:
-            summary = await ctxt.pull_request.render_template(
-                self.config["summary"], extra_variables
+            summary = await jinja2_utils.render_template(
+                ctxt.pull_request, self.config["summary"], extra_variables
             )
-        except context.RenderTemplateFailure as rmf:
+        except jinja2_utils.RenderTemplateFailure as rmf:
             return check_api.Result(
                 check_api.Conclusion.FAILURE,
                 "Invalid summary template",
