+++ mergify_engine/actions/review.py
@@ -21,6 +21,7 @@ from mergify_engine import actions
 from mergify_engine import check_api
 from mergify_engine import config
 from mergify_engine import context
+from mergify_engine import jinja2_utils
 from mergify_engine import rules
 from mergify_engine import signals
 from mergify_engine.actions import utils as action_utils
@@ -75,10 +76,10 @@ class ReviewAction(actions.Action):
 
         if self.config["message"]:
             try:
-                payload["body"] = await ctxt.pull_request.render_template(
-                    self.config["message"]
+                payload["body"] = await jinja2_utils.render_template(
+                    ctxt.pull_request, self.config["message"]
                 )
-            except context.RenderTemplateFailure as rmf:
+            except jinja2_utils.RenderTemplateFailure as rmf:
                 return check_api.Result(
                     check_api.Conclusion.FAILURE, "Invalid review message", str(rmf)
                 )
